proxy DiscoveryHeartbeatEvent's from parent to child be publishing a EurekaHeartbeatEvent and having zuul listen for EurekaHeartbeatEvent as well.
fixes gh-171

+++ spring-cloud-netflix-core/src/main/java/org/springframework/cloud/netflix/eureka/EurekaDiscoveryClientConfiguration.java
@@ -94,11 +94,14 @@ public class EurekaDiscoveryClientConfiguration implements SmartLifecycle, Order
 		if (jerseyClientField != null) {
 			try {
 				jerseyClientField.setAccessible(true);
-				Object obj = jerseyClientField.get(DiscoveryManager.getInstance()
-						.getDiscoveryClient());
-				if (obj != null) {
-					EurekaJerseyClient.JerseyClient jerseyClient = (EurekaJerseyClient.JerseyClient) obj;
-					jerseyClient.destroyResources();
+				if (DiscoveryManager.getInstance() != null &&
+						DiscoveryManager.getInstance().getDiscoveryClient() != null) {
+					Object obj = jerseyClientField.get(DiscoveryManager.getInstance()
+							.getDiscoveryClient());
+					if (obj != null) {
+						EurekaJerseyClient.JerseyClient jerseyClient = (EurekaJerseyClient.JerseyClient) obj;
+						jerseyClient.destroyResources();
+					}
 				}
 			}
 			catch (Exception ex) {
